fix(php8.3): cast usleep() argument to int to silence float-to-int deprecation
PHP 8.3 deprecates implicit float-to-int conversion; usleep() requires int.

+++ Command/RunCommand.php
@@ -165,7 +165,7 @@ class RunCommand extends Command
             $this->startJobs($workerName, $idleTime, $maxJobs, $restrictedQueues, $queueOptionsDefaults, $queueOptions);
 
             $waitTimeInMs = random_int(500, 1000);
-            usleep($waitTimeInMs * 1E3);
+            usleep((int) ($waitTimeInMs * 1E3));
         }
 
         if ($this->verbose) {
