+++ Command/ScheduleCommand.php
@@ -70,7 +70,7 @@ class ScheduleCommand extends Command
 
             $timeToWait = microtime(true) - $lastRunAt + $minJobInterval;
             if ($timeToWait > 0) {
-                usleep($timeToWait * 1E6);
+                usleep((int) ($timeToWait * 1E6));
             }
         }
